webapp/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
Flask-Script
Flask-Bcrypt
Flask-WTF
Jinja2
Werkzeug
requests
PyYAML